refactor(ui): split QuestionPromptSection pure logic and resolved card
- chat/questionAnswers.ts: answer-matrix builders, ready gating, resolved
  summary formatting, option letters (+ unit tests for the sentinel logic)
- chat/questionPromptFocus.ts: composer refocus helper; the
  PermissionPromptSection import now points here
- chat/QuestionResolvedCard.tsx: collapsed resolved summary card
The four active option-row variants stay inline by design (their input
semantics differ). Pure move, markup unchanged.
